Ajustado alteração Estoque

+++ tc/src/main/java/br/com/rdevs/tc/model/dto/DocumentoItemDTO.java
@@ -11,7 +11,7 @@ public class DocumentoItemDTO {
 
     private ProdutoDTO produto;
 
-    private Long qtItem;
+    private Integer qtItem;
 
     private Double valorItem;
